Bus Pirate buffer management revamp
The buffer management of the Bus Pirate driver has been revamped to use
grow-only buffers with a reasonable initial default size so realloc()
will not have to be called in normal operation. A side effect is the
ability to switch to a static buffer without major hassle.
Handle OOM gracefully.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Patrick Georgi <patrick@georgi-clan.de>


git-svn-id: svn://coreboot.org/flashrom/trunk@1541 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1